의자,문 기능 및 사다리 기능 분리
